vdev Knot

Supervises and manages a node.

A node is composed of different parts:

Two simple one-for-one supervisors. One for the connectors and one for the clients.

A logic process which coordinates everything (See Knot.Logic).

A listener which role is to handle incoming connections and be the acceptor (see Knot.Listener).

A supervisor in charge of the client processes (See Knot.Client).

Types

clients()

clients() :: Knot.Via.t | pid

connectors()

connectors() :: Knot.Via.t | pid

socket()

socket() :: :gen_tcp.socket

t()

t() :: Knot.Via.t

Functions

make_handle(uri_or_address)

make_handle(Knot.Via.uri_or_address) :: Knot.Handle.t

start(uri_or_address, block)

start(Knot.Via.uri_or_address, Block.t) :: Knot.Handle.t

start_link(uri_or_address, block)

start_link(Knot.Via.uri_or_address, Block.t) :: {:ok, pid}