vdev Knot.Via

Helps building Registry based process names.

Types

id()

id() :: {String.t, pos_integer, service_type}

service_type()

service_type() :: :node | :clients | :connectors | :logic | :listener

t()

t() :: {:via, Registry, {Knot.Registry, Knot.Via.id}}

uri_or_address()

uri_or_address() :: URI.t | String.t

Functions

clients(uri_or_address)

clients(uri_or_address) :: Knot.clients

Builds a via-tuple for a node clients supervisor given an URI.

Examples

iex> Knot.Via.clients "tcp://localhost:4001"
{:via, Registry, {Knot.Registry, {"localhost", 4001, :clients}}}

connectors(uri_or_address)

connectors(uri_or_address) :: Knot.connectors

Builds a via-tuple for a node connectors supervisor given an URI.

Examples

iex> Knot.Via.connectors "tcp://localhost:4001"
{:via, Registry, {Knot.Registry, {"localhost", 4001, :connectors}}}

id(uri_or_address, suffix)

id(uri_or_address, service_type) :: id

Generates a per-uri unique ID that is suitable for via-tuple generation.

Examples

iex> Knot.Via.id "tcp://localhost:4001", :whatever
{"localhost", 4001, :whatever}

iex> "tcp://localhost:4001" |> URI.parse |> Knot.Via.id(:whatever)
{"localhost", 4001, :whatever}

listener(uri_or_address)

listener(uri_or_address) :: Knot.Listener.t

Builds a via-tuple for a node listener given an URI.

Examples

iex> Knot.Via.listener "tcp://localhost:4001"
{:via, Registry, {Knot.Registry, {"localhost", 4001, :listener}}}

logic(uri_or_address)

logic(uri_or_address) :: Knot.Logic.t

Builds a via-tuple for a node logic given an URI.

Examples

iex> Knot.Via.logic "tcp://localhost:4001"
{:via, Registry, {Knot.Registry, {"localhost", 4001, :logic}}}

make(uri_or_address, suffix)

make(uri_or_address, service_type) :: Knot.Via.t

Given an URI and a suffix, returns a via-tuple compatible with Registry.

Examples

iex> Knot.Via.make "tcp://localhost:4001", :whatever
{:via, Registry, {Knot.Registry, {"localhost", 4001, :whatever}}}

iex> "tcp://localhost:4001" |> URI.parse |> Knot.Via.make(:whatever)
{:via, Registry, {Knot.Registry, {"localhost", 4001, :whatever}}}

node(uri_or_address)

node(uri_or_address) :: Knot.t

Builds a via-tuple for a node given an URI.

Examples

iex> Knot.Via.node "tcp://localhost:4001"
{:via, Registry, {Knot.Registry, {"localhost", 4001, :node}}}

registry()

registry() :: Knot.Registry

to_string(uri_or_address)

to_string(uri_or_address) :: String.t

Transforms a URI into a displayable string.

Examples

iex> Knot.Via.to_string "tcp://localhost:4001"
"localhost:4001"

iex> "tcp://localhost:4001" |> URI.parse |> Knot.Via.to_string
"localhost:4001"